add to headerpanel
