+++ src/content-handlers/iiif/modules/uv-shared-module/HeaderPanelRightOptions.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import HeaderButton from "../uv-shared-module/HeaderButton";
+
+const PagingHeaderPanelRightOptions = () => {
+  return (
+    <>
+      <HeaderButton
+        onClick={() => console.log("Right options button test")}
+        label="Test button"
+      >
+        T{/* icon goes here */}
+      </HeaderButton>
+      <HeaderButton
+        onClick={() => console.log("Left options button test")}
+        label="Test button"
+      >
+        T{/* icon goes here */}
+      </HeaderButton>
+    </>
+  );
+};
+
+export default PagingHeaderPanelRightOptions;
